Android 14.0.0 Release 50 (AP2A.240605.024)

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCZmdzpwAKCRDorT+BmrEO
# eK/kAJ0Qc8hX0IpaFh36xMDaWU5CkqYhyACfTAoS8UiuCbThp5WSk27HBcN9WyA=
# =14Bx
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue Jun 11 00:44:07 2024 EEST
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [marginal]
# gpg: initial-contribution@android.com: Verified 2514 signatures in the past
#      2 years.  Encrypted 4 messages in the past 2 years.
# gpg: WARNING: This key is not certified with sufficiently trusted signatures!
# gpg:          It is not certain that the signature belongs to the owner.
# Primary key fingerprint: 4340 D135 70EF 945E 8381  0964 E8AD 3F81 9AB1 0E78

# By Kangping Dong (45) and others
# Via Automerger Merge Worker (1227) and others
* tag 'android-14.0.0_r50': (327 commits)
  Import translations. DO NOT MERGE ANYWHERE
  Add testAllowedUids for devices without FEATURE_TELEPHONY_SUBSCRIPTION
  Fix flakes due to networks obtained via sync APIs
  Import translations. DO NOT MERGE ANYWHERE
  Revert "24Q2: use mainline netbpfload from apex"
  24Q2: use mainline netbpfload from apex
  Fix the ignore_on_* flags for egress tracing.
  Remove unused library visibility
  Connect to IBluetoothFinder and use it
  Update the imports to androidx.test.filters.*
  Move VpnManagerServiceTest and VpnTest to frameworks
  Use a TAP test network for MeshCoP service test cases
  Fix flaky multicast tests in Thread E2E test.
  NetBpfLoader: create /sys/fs/bpf/loader dir
  Revert "Use a TAP test network for MeshCoP service test cases"
  Use a TAP test network for MeshCoP service test cases
  [mdns] add service-side impl for NSD service TTL support
  Adding additional transport satellite support at getSubIdForMobile()
  netbpfload: fail if platform exec of apex fails
  [Thread] make Thread state customisable via resource overlay
  ...

 Conflicts:
	netbpfload/NetBpfLoad.cpp
	service/src/com/android/server/ConnectivityService.java

+++ tests/native/utilities/firewall.cpp
@@ -60,10 +60,10 @@ Result<void> Firewall::addRule(uint32_t uid, UidOwnerMatchType match, uint32_t i
     // iif should be non-zero if and only if match == MATCH_IIF
     if (match == IIF_MATCH && iif == 0) {
         return Errorf("Interface match {} must have nonzero interface index",
-                      static_cast<int>(match));
+                      static_cast<uint32_t>(match));
     } else if (match != IIF_MATCH && iif != 0) {
         return Errorf("Non-interface match {} must have zero interface index",
-                      static_cast<int>(match));
+                      static_cast<uint32_t>(match));
     }
 
     std::lock_guard guard(mMutex);
@@ -71,14 +71,14 @@ Result<void> Firewall::addRule(uint32_t uid, UidOwnerMatchType match, uint32_t i
     if (oldMatch.ok()) {
         UidOwnerValue newMatch = {
                 .iif = iif ? iif : oldMatch.value().iif,
-                .rule = static_cast<uint8_t>(oldMatch.value().rule | match),
+                .rule = oldMatch.value().rule | match,
         };
         auto res = mUidOwnerMap.writeValue(uid, newMatch, BPF_ANY);
         if (!res.ok()) return Errorf("Failed to update rule: {}", res.error().message());
     } else {
         UidOwnerValue newMatch = {
                 .iif = iif,
-                .rule = static_cast<uint8_t>(match),
+                .rule = match,
         };
         auto res = mUidOwnerMap.writeValue(uid, newMatch, BPF_ANY);
         if (!res.ok()) return Errorf("Failed to add rule: {}", res.error().message());
@@ -93,7 +93,7 @@ Result<void> Firewall::removeRule(uint32_t uid, UidOwnerMatchType match) {
 
     UidOwnerValue newMatch = {
             .iif = (match == IIF_MATCH) ? 0 : oldMatch.value().iif,
-            .rule = static_cast<uint8_t>(oldMatch.value().rule & ~match),
+            .rule = oldMatch.value().rule & ~match,
     };
     if (newMatch.rule == 0) {
         auto res = mUidOwnerMap.deleteValue(uid);
